Cleaned up warnings in test

+++ test/list_test.h
@@ -7,18 +7,18 @@ DEFINE_TEST(TEST_LIST_ADD_FIRST)
 
     klist_add_first(int_list, 1);
     CHECK_TEST(klist_size(int_list) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
 
     klist_add_first(int_list, 2);
     CHECK_TEST((int)klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 1);
 
     klist_add_first(int_list, 3);
     CHECK_TEST(klist_size(int_list) == 3);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 3);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 2) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 3);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 2) == 1);
 
     klist_destroy(int_list);
 
@@ -32,18 +32,18 @@ DEFINE_TEST(TEST_LIST_ADD)
 
     klist_add(int_list, 1);
     CHECK_TEST(klist_size(int_list) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
 
     klist_add(int_list, 2);
     CHECK_TEST((int)klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 2);
 
     klist_add(int_list, 3);
     CHECK_TEST(klist_size(int_list) == 3);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 2) == 3);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 2) == 3);
 
     klist_destroy(int_list);
 
@@ -57,18 +57,18 @@ DEFINE_TEST(TEST_LIST_ADD_LAST)
 
     klist_add_last(int_list, 1);
     CHECK_TEST(klist_size(int_list) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
 
     klist_add_last(int_list, 2);
     CHECK_TEST((int)klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 2);
 
     klist_add_last(int_list, 3);
     CHECK_TEST(klist_size(int_list) == 3);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 2) == 3);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 2) == 3);
 
     klist_destroy(int_list);
 
@@ -86,10 +86,10 @@ DEFINE_TEST(TEST_LIST_ADD_AT)
     klist_add_at(int_list, 4, 1);
 
     CHECK_TEST(klist_size(int_list) == 4);
-    CHECK_TEST(klist_get_at(int, int_list, 0) == 1);
-    CHECK_TEST(klist_get_at(int, int_list, 1) == 4);
-    CHECK_TEST(klist_get_at(int, int_list, 2) == 2);
-    CHECK_TEST(klist_get_at(int, int_list, 3) == 3);
+    CHECK_TEST(klist_get_at(size_t, int_list, 0) == 1);
+    CHECK_TEST(klist_get_at(size_t, int_list, 1) == 4);
+    CHECK_TEST(klist_get_at(size_t, int_list, 2) == 2);
+    CHECK_TEST(klist_get_at(size_t, int_list, 3) == 3);
 
     klist_destroy(int_list);
 
@@ -126,8 +126,8 @@ DEFINE_TEST(TEST_LIST_REMOVE)
 
     klist_remove(int_list, 2);
     CHECK_TEST(klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_first(int, int_list) == 1);
-    CHECK_TEST(klist_get_last(int, int_list) == 3);
+    CHECK_TEST(klist_get_first(size_t, int_list) == 1);
+    CHECK_TEST(klist_get_last(size_t, int_list) == 3);
 
     klist_destroy(int_list);
 
@@ -146,8 +146,8 @@ DEFINE_TEST(TEST_LIST_REMOVE_AT)
 
     klist_remove_at(int_list, 1);
     CHECK_TEST(klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_first(int, int_list) == 1);
-    CHECK_TEST(klist_get_last(int, int_list) == 3);
+    CHECK_TEST(klist_get_first(size_t, int_list) == 1);
+    CHECK_TEST(klist_get_last(size_t, int_list) == 3);
 
     klist_destroy(int_list);
 
@@ -166,8 +166,8 @@ DEFINE_TEST(TEST_LIST_REMOVE_LAST)
 
     klist_remove_last(int_list);
     CHECK_TEST(klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_first(int, int_list) == 1);
-    CHECK_TEST(klist_get_last(int, int_list) == 2);
+    CHECK_TEST(klist_get_first(size_t, int_list) == 1);
+    CHECK_TEST(klist_get_last(size_t, int_list) == 2);
 
     klist_destroy(int_list);
 
@@ -186,8 +186,8 @@ DEFINE_TEST(TEST_LIST_REMOVE_FIRST)
 
     klist_remove_first(int_list);
     CHECK_TEST(klist_size(int_list) == 2);
-    CHECK_TEST(klist_get_first(int, int_list) == 2);
-    CHECK_TEST(klist_get_last(int, int_list) == 3);
+    CHECK_TEST(klist_get_first(size_t, int_list) == 2);
+    CHECK_TEST(klist_get_last(size_t, int_list) == 3);
 
     klist_destroy(int_list);
